+++ app/src/main/java/com/b07/planetze/common/Emissions.java
@@ -1,4 +1,4 @@
-package com.b07.planetze.util;
+package com.b07.planetze.common;
 
 import androidx.annotation.NonNull;
