+++ app/src/main/java/com/b07/planetze/common/User.java
@@ -1,4 +1,4 @@
-package com.b07.planetze.database;
+package com.b07.planetze.common;
 
 /**
  * A user.
