+++ app/src/main/java/com/b07/planetze/database/FakeDatabase.java
@@ -1,9 +1,10 @@
 package com.b07.planetze.database;
 
-import com.b07.planetze.util.DateInterval;
-import com.b07.planetze.util.Emissions;
-import com.b07.planetze.util.Mass;
-import com.b07.planetze.util.Result;
+import com.b07.planetze.common.DateInterval;
+import com.b07.planetze.common.DatedEmissions;
+import com.b07.planetze.common.Emissions;
+import com.b07.planetze.common.UserId;
+import com.b07.planetze.type.Result;
 
 import java.time.LocalDate;
 import java.util.ArrayList;
